+++ source/bronze/tests/infrastructure/streams/test_kafka_stream.py
@@ -4,7 +4,7 @@ import tests.helpers.environment_variables_helpers as environment_variables_help
 from opengeh_bronze.infrastructure.streams.kafka_stream import KafkaStream
 
 
-@mock.patch("opengeh_bronze.infrastructure.streams.kafka_stream.shared_helpers.get_checkpoint_path")
+@mock.patch("opengeh_bronze.infrastructure.streams.kafka_stream.path_helper.get_checkpoint_path")
 def test__submit_transactions__should_set_environment_variables(mock_get_checkpoint_path):
     # Arrange
     mock_get_checkpoint_path.return_value = "/tmp/checkpoints/start_write_stream_test"
@@ -18,7 +18,7 @@ def test__submit_transactions__should_set_environment_variables(mock_get_checkpo
     kafka_stream.submit_transactions(spark_mock)
 
 
-@mock.patch("opengeh_bronze.infrastructure.streams.kafka_stream.shared_helpers.get_checkpoint_path")
+@mock.patch("opengeh_bronze.infrastructure.streams.kafka_stream.path_helper.get_checkpoint_path")
 def test__write_stream__should_set_environment_variables(mock_get_checkpoint_path):
     # Arrange
     mock_get_checkpoint_path.return_value = "/tmp/checkpoints/start_write_stream_test"
